#14 Adapt javadoc and order of method-usage within a type hierarchy
