+++ processor/src/main/java/org/mapstruct/ap/util/Collections.java
@@ -85,4 +85,15 @@ public class Collections {
 
         return result;
     }
+
+    public static <E> boolean hasNonNullElements(Iterable<E> elements) {
+        if ( elements != null ) {
+            for ( E e : elements ) {
+                if ( e != null ) {
+                    return true;
+                }
+            }
+        }
+        return false;
+    }
 }
